+++ src/pages/Content/ControlBar copy.tsx
@@ -78,7 +78,7 @@ const AIPanel = () => (
           marginBottom: '0.5em',
         }}
       >
-        <CTFlowAI actions={[]} onBack={() => {}} />
+        <CTFlowAI />
       </div>
     </div>
   </div>
